Module Make.Local_config

The configuration of a node to be run as a local node i.e., one that can not send or receive messages with other nodes.

type t = {

node_name : string;
(*
The name of this node.
*)

}